requirements.txt
python-docx
PyPDF2
python-pptx
rich
